requirements-server.txt
absl-py==2.1.0
accelerate==0.31.0
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
APScheduler==3.10.4
arrow==1.3.0
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
av==9.2.0
Babel==2.15.0
backoff==1.11.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.3
bibtexparser==2.0.0b7
binaryornot==0.4.4
bitsandbytes==0.43.1
black==24.4.2
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
clldutils==3.22.2
cloudpickle==3.0.0
codecarbon==1.2.0
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.8.2
contourpy==1.1.1
cookiecutter==1.7.3
csvw==3.3.0
cycler==0.12.1
dash==2.17.0
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
datasets==2.19.2
decorator==5.1.1
decord==0.6.0
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@9ba16d6d9252fd4d820fd1317a41af6bd9aefd23
dill==0.3.4
dlinfo==1.2.1
docker-pycreds==0.4.0
einops==0.8.0
evaluate==0.4.2
exceptiongroup==1.2.1
execnet==2.1.1
faiss-cpu==1.8.0
filelock==3.14.0
fire==0.6.0
flash-attn==2.5.9.post1
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.3.1
fugashi==1.3.2
fvcore==0.1.5.post20221221
gitdb==4.0.11
GitPython==3.1.18
google-auth==2.30.0
google-auth-oauthlib==1.0.0
greenlet==3.0.3
grpcio==1.64.1
huggingface-hub==0.23.3
humanfriendly==10.0
hydra-core==1.3.2
hypothesis==6.103.1
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
iniconfig==2.0.0
iopath==0.1.9
ipadic==1.0.0
isodate==0.6.1
isort==5.13.2
itsdangerous==2.0.1
Jinja2==3.1.4
jinja2-time==0.2.0
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kenlm==0.2.0
kiwisolver==1.4.5
language-tags==1.2.0
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.41.1
lxml==5.2.2
Mako==1.3.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.12.2
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.1
nltk==3.8.1
numba==0.58.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.16.1
onnxruntime==1.18.0
onnxruntime-tools==1.7.0
optuna==3.6.1
packaging==24.1
pandas==2.0.3
parameterized==0.9.0
pathspec==0.12.1
peft==0.12.0
phonemizer==3.2.1
pillow==10.3.0
pkgutil_resolve_name==1.3.10
plac==1.4.3
platformdirs==4.2.2
plotly==5.22.0
pluggy==1.5.0
pooch==1.8.2
portalocker==2.0.0
poyo==0.5.0
protobuf==5.27.1
psutil==5.9.8
py-cpuinfo==9.0.0
py3nvml==0.2.7
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycocotools==2.0.7
pycparser==2.22
pyctcdecode==0.5.0
pydantic==2.7.3
pydantic_core==2.18.4
Pygments==2.18.0
pygtrie==2.5.0
pylatexenc==2.10
pynvml==11.5.0
pyparsing==3.1.2
pypng==0.20220715.0
pytesseract==0.3.10
pytest==7.4.4
pytest-rich==0.1.1
pytest-timeout==2.3.1
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
ray==2.10.0
rdflib==7.0.0
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
retrying==1.3.4
rfc3986==1.5.0
rhoknp==1.3.0
rich==13.7.1
rjieba==0.1.11
rouge-score==0.1.2
rpds-py==0.18.1
rsa==4.9
ruff==0.4.4
sacrebleu==1.5.1
sacremoses==0.1.1
safetensors==0.4.3
scikit-learn==1.3.2
scipy==1.10.1
segments==2.2.1
sentencepiece==0.2.0
sentry-sdk==2.5.1
setproctitle==1.3.3
sigopt==8.8.3
six==1.16.0
smmap==5.0.1
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
SQLAlchemy==2.0.30
SudachiDict-core==20240409
SudachiPy==0.6.8
sympy==1.12.1
tabulate==0.9.0
tenacity==8.3.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.4.0
text-unidecode==1.3
threadpoolctl==3.5.0
timeout-decorator==0.5.0
timm==0.9.16
tokenizers==0.19.1
tomli==2.0.1
torch==2.3.1+cu121
torchaudio==2.3.1+cu121
torchvision==0.18.1+cu121
tqdm==4.66.4
-e git+https://github.com/huggingface/transformers@dcdda5324bcc7a750b5e40e11dd795442204ff27#egg=transformers
triton==2.3.1
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
unidic==1.1.0
unidic-lite==1.0.8
uritemplate==4.1.1
urllib3==1.26.18
wandb==0.17.1
wasabi==0.10.1
Werkzeug==3.0.3
xmltodict==0.13.0
xxhash==3.4.1
yacs==0.1.8
yarl==1.9.4
zipp==3.19.2
